requirements.txt
python-telegram-bot==20.3
speedtest-cli==2.1.3
